Rename teacher.ts and teacher.test.ts for teacherSubject; Added token user email and teacher email validation in teacherSubjects.ts; Modified serverless.yaml path for the endpoint teacherSubjects
