+++ serverless.yaml
@@ -135,23 +135,21 @@ functions:
             path: /homework/{homeworkId?}
             method: get
             cors: true
-  teacher:
-      name: ${self:custom.base}-teacher
-      description: 'teacher function'
-      handler: src/handlers/subjects/teacher.handler
+  teacherSubjects:
+      name: ${self:custom.base}-teacherSubjects
+      description: 'teacherSubjects function'
+      handler: src/handlers/teacherSubjects.handler
       timeout: 30
       memorySize: 264
-      environment:
-        TOKEN_SECRET: 'PARANGARICUTIRIMICUARO'
       package:
         patterns:
           - '!src/**'
-          - 'src/handlers//subjects/teacher.*'
+          - 'src/handlers/teacherSubjects.*'
           - 'src/constants.*'
           - 'src/dynamodb/*'
           - 'src/helpers/**/*'
       events:
         - http:
-            path: /subjects/teacher/{teacherEmail?}
+            path: /teacherSubjects/{teacherEmail?}
             method: get
             cors: true
